update spot price datatype

+++ backend/db/migrations/20240829233627-create-spot.js
@@ -54,7 +54,7 @@ module.exports = {
           allowNull: false,
         },
         price: {
-          type: Sequelize.DECIMAL,
+          type: Sequelize.FLOAT,
           allowNull: false,
         },
         createdAt: {
